requirements.txt
blinker==1.7.0
casregnum==1.0.1
chemformula==1.3.1
click==8.1.7
Flask==3.0.3
Flask-Mail==0.9.1
itsdangerous==2.2.0
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
numpy==1.26.4
pillow==10.3.0
python-docx==1.1.0
typing_extensions==4.10.0
Werkzeug==3.0.2
